Refactoring with ticket ID
