+++ tests/Tests/ORM/Functional/Ticket/DDC11871/DDC11871Order.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\Tests\ORM\Functional\Ticket\SkipPrivateSetProperties;
+namespace Doctrine\Tests\ORM\Functional\Ticket\DDC11871;
 
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity]
-#[ORM\Table(name: 'Test_Class_With_Private_Set_Properties_Order')]
-class OrderPrivateSetProperties
+#[ORM\Table(name: 'DDC11871_Order')]
+class DDC11871Order
 {
     #[ORM\Id]
     #[ORM\Column(type: 'integer')]
@@ -19,8 +19,8 @@ class OrderPrivateSetProperties
     private(set) string $company;
 
     public function __construct(
-        #[ORM\ManyToOne(targetEntity: UserPrivateSetProperties::class, fetch: 'LAZY')]
-        private(set) UserPrivateSetProperties $user,
+        #[ORM\ManyToOne(targetEntity: DDC11871User::class, fetch: 'LAZY')]
+        private(set) DDC11871User $user,
     ) {
         $this->company = $user->company;
     }
