+++ tests/Tests/ORM/Functional/Ticket/DDC11871/DDC11871Test.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\Tests\ORM\Functional\Ticket\SkipPrivateSetProperties;
+namespace Doctrine\Tests\ORM\Functional\Ticket\DDC11871;
 
 use Doctrine\Tests\OrmFunctionalTestCase;
 
 use const PHP_VERSION_ID;
 
-class ClassWithPrivateSetPropertiesTest extends OrmFunctionalTestCase
+class DDC11871Test extends OrmFunctionalTestCase
 {
     public function setUp(): void
     {
@@ -19,21 +19,21 @@ class ClassWithPrivateSetPropertiesTest extends OrmFunctionalTestCase
         }
 
         $this->setUpEntitySchema([
-            UserPrivateSetProperties::class,
-            OrderPrivateSetProperties::class,
+            DDC11871User::class,
+            DDC11871Order::class,
         ]);
     }
 
     public function testEntityHydratation(): void
     {
-        $user = new UserPrivateSetProperties('Some company');
+        $user = new DDC11871User('Some company');
         $this->_em->persist($user);
-        $order = new OrderPrivateSetProperties($user);
+        $order = new DDC11871Order($user);
         $this->_em->persist($order);
         $this->_em->flush();
         $this->_em->clear();
 
-        $hydrated = $this->_em->getRepository(OrderPrivateSetProperties::class)->findAll();
+        $hydrated = $this->_em->getRepository(DDC11871Order::class)->findAll();
         self::assertCount(1, $hydrated);
     }
 }
